# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sales Support API",
        "description": "SES企業向け営業支援システム API",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "ローカル開発環境"
        }
    ],
    "paths": {
...
        "/api/v1/matching/projects/{id}/engineers": {
            "get": {
                "tags": [
                    "Matching"
                ],
                "summary": "案件に対するおすすめ技術者一覧",
                "operationId": "c32b81eb71c8d9ac2f685e8c2a9ccf63",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "案件ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "スコア付き技術者一覧"
                    },
                    "404": {
                        "description": "案件が見つかりません"
                    },
                    "401": {
                        "description": "認証エラー"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/matching/engineers/{id}/projects": {
            "get": {
                "tags": [
                    "Matching"
                ],
                "summary": "技術者に対するおすすめ案件一覧",
                "operationId": "52bef4cc70d81e25a46182b3b4173d39",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "技術者ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "スコア付き案件一覧"
                    },
                    "404": {
                        "description": "技術者が見つかりません"
                    },
                    "401": {
                        "description": "認証エラー"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/matching/projects/{projectId}/engineers/{engineerId}": {
            "get": {
                "tags": [
                    "Matching"
                ],
                "summary": "案件×技術者マッチングスコア詳細",
                "operationId": "d64bbf5db4e98439dde49021a6c1f5f1",
                "parameters": [
                    {
                        "name": "projectId",
                        "in": "path",
                        "description": "案件ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "engineerId",
                        "in": "path",
                        "description": "技術者ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "スコア詳細とAI説明文"
                    },
                    "401": {
                        "description": "認証エラー"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/matching/skills": {
            "get": {
                "tags": [
                    "Matching"
                ],
                "summary": "スキルマスタ一覧取得",
                "operationId": "a4bae7cb9fedbb2da75c188bd2cfb9ec",
                "parameters": [
                    {
                        "name": "category",
                        "in": "query",
                        "description": "カテゴリ",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "language",
                                "framework",
                                "database",
                                "infrastructure",
                                "other"
                            ]
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "スキル名で検索",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "スキル一覧"
                    },
                    "401": {
                        "description": "認証エラー"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Matching"
                ],
                "summary": "スキル登録（管理用）",
                "operationId": "05640a0b0626bb7012bf0ecaf5b49d1c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Laravel"
                                    },
                                    "category": {
                                        "type": "string",
                                        "enum": [
                                            "language",
                                            "framework",
                                            "database",
                                            "infrastructure",
                                            "other"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "登録成功"
                    },
                    "422": {
                        "description": "バリデーションエラー"
                    },
                    "401": {
                        "description": "認証エラー"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Supabase Auth JWT トークン",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}